Check if the file exists before moving it.

+++ trunk/igesture-framework/src/main/java/org/ximtec/igesture/util/ZipFS.java
@@ -244,6 +244,8 @@ public class ZipFS {
          zipFile.close();
          file.delete();
       }
-      FileUtils.moveFile(tmpFile, file);
+      if(!file.exists()){
+         FileUtils.moveFile(tmpFile, file);
+      }
    }
 }
